async_api/requirements.txt
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
Authlib==1.3.1
certifi==2024.2.2
cffi==1.17.0
click==8.1.7
cryptography==43.0.0
Deprecated==1.2.14
elasticsearch==7.9.1
fastapi==0.101.0
googleapis-common-protos==1.59.1
grpcio==1.65.5
gunicorn==21.2.0
h11==0.14.0
importlib_metadata==8.0.0
opentelemetry-api==1.26.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.9.15
packaging==24.1
protobuf==4.25.4
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.9.0
python-dotenv==1.0.1
redis==4.4.2
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
thrift==0.20.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
wrapt==1.16.0
zipp==3.20.0

aiokafka==0.11.0
async_timeout==4.0.3
packaging==24.1
typing_extensions==4.12.2

aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
yarl==1.11.1